Update to 147.7703.
Update to commit 6684a0e296af9051b7e08b8dc149e4f295ab56f8

+++ gpu/ipc/service/webgpu_command_buffer_stub.cc
@@ -55,7 +55,9 @@ WebGPUCommandBufferStub::WebGPUCommandBufferStub(
 
 WebGPUCommandBufferStub::~WebGPUCommandBufferStub() {
   // Must run before memory_tracker_ is destroyed.
-  decoder_context()->Destroy(false);
+  if (decoder_context()) {
+    decoder_context()->Destroy(false);
+  }
 
   memory_tracker_ = nullptr;
 }
@@ -136,10 +138,6 @@ gpu::ContextResult WebGPUCommandBufferStub::Initialize(
   initialized_ = true;
   return gpu::ContextResult::kSuccess;
 #endif  // BUILDFLAG(IS_FUCHSIA)
-}
-
-MemoryTracker* WebGPUCommandBufferStub::GetContextGroupMemoryTracker() const {
-  return nullptr;
 }
 
 base::WeakPtr<CommandBufferStub> WebGPUCommandBufferStub::AsWeakPtr() {
